fix(bitvec): change popcount_words return type from u32 to usize to prevent overflow
The `popcount_words` function and all its SIMD/platform-specific variants
previously returned `u32`, which overflows for inputs exceeding ~512MB of
all-ones data (67M words × 64 bits = 4.29B > u32::MAX).

- Change return type of `popcount_words` and all internal implementations
  (default, portable, NEON, AVX-512, x86 scalar) from `u32` to `usize`
- Update accumulator variables from `0u32` to `0usize` in all variants
- Cast per-word counts to `usize` before accumulation to avoid intermediate
  overflow in NEON and AVX-512 paths
- Remove redundant `as usize` cast at call site in `bitvec.rs` since return
  type now matches usage directly
- Add comprehensive regression and boundary tests including
  `test_popcount_words_exceeds_u32_max` which documents the overflow scenario,
  chunk boundary tests for SIMD loop edges, all-ones/all-zeros tests, and
  mixed-density pattern tests

+++ src/bits/bitvec.rs
@@ -87,7 +87,7 @@ impl BitVec {
         }
 
         // Count total ones using the selected popcount implementation
-        let ones_count: usize = popcount_words(&words) as usize;
+        let ones_count: usize = popcount_words(&words);
 
         // Build rank directory
         let rank_dir = RankDirectory::build(&words);
